requirements.txt
about-time==4.2.1
alive-progress==3.1.4
ansicolors==1.1.8
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.1.0
awsipranges==0.3.3
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.13.0
azure-mgmt-authorization==3.0.0
azure-mgmt-core==1.4.0
azure-mgmt-security==5.0.0
azure-mgmt-storage==21.0.0
azure-mgmt-subscription==3.1.1
azure-storage-blob==12.16.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
boto3==1.5.32
botocore==1.8.50
cachetools==5.3.1
click==8.1.3
click-plugins==1.1.1
cloudtracker==2.1.5
colorama==0.4.6
comm==0.1.2
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
detect-secrets==1.4.0
distlib==0.3.7
dnspython==2.3.0
docutils==0.20.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.2
flask==2.3.2
fonttools==4.38.0
fqdn==1.5.1
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.22.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
grapheme==0.6.0
gunicorn==20.1.0
httplib2==0.22.0
python-dotenv==1.0.1
importlib-metadata==5.1.0
iniconfig==1.1.1
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
jmespath==0.9.3
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.8
jupyter-console==6.4.4
jupyter-core==5.1.0
jupyter-events==0.5.0
jupyter-server==2.0.1
jupyter-server-terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
markupsafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
msal==1.22.0
msal-extensions==1.0.0
msgraph-core==0.2.2
msrest==0.7.1
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook-shim==0.2.2
# numpy==1.23.5
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.3.0
pip==23.2
platformdirs==3.9.1
portalocker==2.7.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.23.4
# prowler==3.7.1
pyzmq==25.1.2
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pygments==2.13.0
pyjwt==2.8.0
pymongo==4.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
pyyaml==4.2b4
qtconsole==5.4.0
qtpy==2.3.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
s3transfer==0.1.13
scapy==2.5.0
schema==0.7.5
# scikit-learn==1.2.0
# scipy==1.9.3
seaborn==0.12.1
send2trash==1.8.0
shodan==1.29.1
slack-sdk==3.21.3
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
typing-extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
virtualenv==20.24.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
werkzeug==2.3.6
widgetsnbextension==4.0.4
xlsxwriter==3.1.2
zipp==3.11.0

